Move test generated clients into a buildable crate
