+++ .gitignore
@@ -1,1 +1,3 @@
 target/
+
+progenitor-impl/tests/output/Cargo.lock
