+++ progenitor-impl/tests/output/src/buildomat_builder_tagged.rs
