+++ progenitor-impl/tests/output/src/buildomat_cli.rs
@@ -1,4 +1,4 @@
-use sdk::*;
+use crate::buildomat_builder::*;
 pub struct Cli<T: CliOverride = ()> {
     client: Client,
     over: T,
