+++ progenitor-impl/tests/output/src/buildomat_httpmock.rs
@@ -3,7 +3,7 @@ pub mod operations {
     #![doc = r" wrappers for each operation. Each can be converted to"]
     #![doc = r" its inner type with a call to `into_inner()`. This can"]
     #![doc = r" be used to explicitly deviate from permitted values."]
-    use sdk::*;
+    use crate::buildomat_builder::*;
     pub struct ControlHoldWhen(httpmock::When);
     impl ControlHoldWhen {
         pub fn new(inner: httpmock::When) -> Self {
