+++ progenitor-impl/tests/output/src/keeper_cli.rs
@@ -1,4 +1,4 @@
-use sdk::*;
+use crate::keeper_builder::*;
 pub struct Cli<T: CliOverride = ()> {
     client: Client,
     over: T,
