+++ progenitor-impl/tests/output/src/nexus_builder_tagged.rs
