+++ progenitor-impl/tests/output/src/nexus_cli.rs
@@ -1,4 +1,4 @@
-use sdk::*;
+use crate::nexus_builder::*;
 pub struct Cli<T: CliOverride = ()> {
     client: Client,
     over: T,
