+++ progenitor-impl/tests/output/src/param_collision_cli.rs
@@ -1,4 +1,4 @@
-use sdk::*;
+use crate::param_collision_builder::*;
 pub struct Cli<T: CliOverride = ()> {
     client: Client,
     over: T,
