+++ progenitor-impl/tests/output/src/param_overrides_cli.rs
@@ -1,4 +1,4 @@
-use sdk::*;
+use crate::param_overrides_builder::*;
 pub struct Cli<T: CliOverride = ()> {
     client: Client,
     over: T,
